custom hook for guiderLabel functions created label function removed from asid optional rendering for announcement labels
