+++ muikku-frontend/muikkuV3/components/guider/body/application/aside/students.tsx
@@ -13,17 +13,6 @@ import { useTranslation } from "react-i18next";
 import useIsAtBreakpoint from "~/hooks/useIsAtBreakpoint";
 import { breakpoints } from "~/util/breakpoints";
 
-import {
-  Collaborator,
-  GenericTag,
-} from "~/components/general/tag-update-dialog";
-import { displayNotification } from "~/actions/base/notifications";
-import MApi, { isMApiError } from "~/api/api";
-
-import {
-  updateGuiderFilterLabel,
-  removeGuiderFilterLabel,
-} from "~/actions/main-function/guider";
 import GuiderLabel from "./students/label";
 
 /**
@@ -35,170 +24,6 @@ const StudentNavigationAside = () => {
 
   const { t } = useTranslation(["flags"]);
   const isMobileWidth = useIsAtBreakpoint(breakpoints.breakpointPad);
-  const dispatch = useDispatch();
-
-  /**
-   * Handles delete category
-   * @param tag tag to be deleted
-   */
-  const handleDelete = (tag: GenericTag) => {
-    dispatch(removeGuiderFilterLabel({ id: tag.id }));
-  };
-
-  /**
-   * Handles update category
-   * @param tag tag to be updated
-   * @param onSuccess optional success callback
-   * @param onFail optional fail callback
-   */
-  const handleUpdate = (
-    tag: GenericTag,
-    onSuccess?: () => void,
-    onFail?: () => void
-  ) => {
-    if (tag.collaborators.toAdd.length > 0) {
-      handleAddCollaborators(tag);
-    }
-    if (tag.collaborators.toRemove.length > 0) {
-      handleRemoveCollaborators(tag);
-    }
-    dispatch(
-      updateGuiderFilterLabel({
-        id: tag.id,
-        name: tag.label,
-        color: "#" + tag.color.toString(16).padStart(6, "0"),
-        description: tag.description || "",
-        ownerIdentifier: tag.ownerIdentifier,
-        success: onSuccess,
-        fail: onFail,
-      })
-    );
-  };
-
-  /**
-   * Handles adding collaborators to a tag
-   * @param tag tag
-   * @param success success
-   * @param fail fail
-   */
-  const handleAddCollaborators = async (
-    tag: GenericTag,
-    success?: () => void,
-    fail?: () => void
-  ) => {
-    const flagId = tag.id;
-    const userApi = MApi.getUserApi();
-
-    try {
-      for (const collaborator of tag.collaborators.toAdd) {
-        await userApi.createFlagShare({
-          flagId,
-          createFlagShareRequest: {
-            flagId,
-            userIdentifier: collaborator.identifier,
-          },
-        });
-      }
-
-      success && success();
-      dispatch(
-        displayNotification(
-          t("notifications.addCollaboratorSuccess", {
-            ns: "flags",
-            collaboratorCount: tag.collaborators.toAdd.length,
-          }),
-          "success"
-        )
-      );
-    } catch (e) {
-      if (!isMApiError(e)) {
-        throw e;
-      }
-      fail && fail();
-      dispatch(
-        displayNotification(
-          t("notifications.addCollaboratorError", {
-            ns: "flags",
-            error: e.message,
-          }),
-          "error"
-        )
-      );
-    }
-  };
-
-  /**
-   * Handles removing collaborators from a tag
-   * @param tag tag
-   * @param success success
-   * @param fail fail
-   */
-  const handleRemoveCollaborators = async (
-    tag: GenericTag,
-    success?: () => void,
-    fail?: () => void
-  ) => {
-    const flagId = tag.id;
-    const userApi = MApi.getUserApi();
-
-    try {
-      let collaborators: Collaborator[] = [];
-      const flagShares = await userApi.getFlagShares({
-        flagId,
-      });
-
-      // If no collaborators are given, remove all collaborators
-
-      if (!tag.collaborators || tag.collaborators.toRemove.length === 0) {
-        collaborators = flagShares.map((share) => ({
-          id: share.id,
-          identifier: share.userIdentifier,
-        }));
-      } else {
-        collaborators = flagShares
-          .filter((share) =>
-            tag.collaborators.toRemove.find(
-              (c) => c.id === share.user.userEntityId
-            )
-          )
-          .map((share) => ({
-            id: share.id,
-            identifier: share.userIdentifier,
-          }));
-      }
-      for (const collaborator of collaborators) {
-        await userApi.deleteFlagShare({
-          flagId,
-          shareId: collaborator.id,
-        });
-      }
-
-      success && success();
-      dispatch(
-        displayNotification(
-          t("notifications.removeCollaboratorSuccess", {
-            ns: "flags",
-            collaboratorCount: collaborators.length,
-          }),
-          "success"
-        )
-      );
-    } catch (e) {
-      if (!isMApiError(e)) {
-        throw e;
-      }
-      fail && fail();
-      dispatch(
-        displayNotification(
-          t("notifications.removeCollaboratorError", {
-            ns: "flags",
-            error: e.message,
-          }),
-          "error"
-        )
-      );
-    }
-  };
 
   const locationData = queryString.parse(
     document.location.hash.split("?")[1] || "",
@@ -255,9 +80,6 @@ const StudentNavigationAside = () => {
                         { arrayFormat: "bracket" }
                       )
                 }
-                onRemoveCollaborators={handleRemoveCollaborators}
-                onDelete={handleDelete}
-                onUpdate={handleUpdate}
               />
             );
           })}
