+++ muikku-frontend/muikkuV3/components/guider/body/application/aside/students/label.tsx
@@ -1,18 +1,16 @@
 import * as React from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { displayNotification } from "~/actions/base/notifications";
-import { hexToColorInt } from "~/util/modifiers";
-import MApi, { isMApiError } from "~/api/api";
+import { useSelector } from "react-redux";
 import { StateType } from "~/reducers";
-import { useTranslation } from "react-i18next";
+
 import {
   NavigationElement,
   NavigationDropdown,
   DropdownWrapperProps,
 } from "~/components/general/navigation";
-import { GenericTag } from "~/components/general/tag-update-dialog";
-import { UserFlag } from "~/generated/client";
 
+import { UserFlag } from "~/generated/client";
+import { useGuiderLabel } from "./hooks/useGuiderLabel";
+import { useTranslation } from "react-i18next";
 /**
  * Props for GuiderLabel
  */
@@ -20,13 +18,6 @@ interface GuiderLabelProps {
   label: UserFlag;
   isActive?: boolean;
   hash?: string;
-  onDelete: (tag: GenericTag, success?: () => void, fail?: () => void) => void;
-  onUpdate: (tag: GenericTag, success?: () => void, fail?: () => void) => void;
-  onRemoveCollaborators: (
-    tag: GenericTag,
-    success?: () => void,
-    fail?: () => void
-  ) => void;
 }
 
 /**
@@ -35,59 +26,11 @@ interface GuiderLabelProps {
  * @returns JSX.Element
  */
 const GuiderLabel: React.FC<GuiderLabelProps> = (props: GuiderLabelProps) => {
-  const { label, onDelete, onUpdate, onRemoveCollaborators, isActive, hash } =
-    props;
+  const { label, isActive, hash } = props;
   const { status } = useSelector((state: StateType) => state);
-  const [tag, setTag] = React.useState<GenericTag>({
-    id: label.id,
-    label: label.name,
-    color: hexToColorInt(label.color),
-    description: label.description || "",
-    ownerIdentifier: label.ownerIdentifier,
-    collaborators: { all: [], toAdd: [], toRemove: [] },
-  });
   const { t } = useTranslation(["flags"]);
-  const dispatch = useDispatch();
-
-  /**
-   * Load collaborators
-   */
-  const loadCollaborators = React.useCallback(
-    async (tagId: number) => {
-      const userApi = MApi.getUserApi();
-      try {
-        const sharesResult = await userApi.getFlagShares({
-          flagId: tagId,
-        });
-
-        setTag((prevTag) => ({
-          ...prevTag,
-          collaborators: { all: sharesResult, toRemove: [], toAdd: [] },
-        }));
-      } catch (e) {
-        if (!isMApiError(e)) {
-          throw e;
-        }
-        dispatch(displayNotification(e.message, "error"));
-      }
-    },
-    [dispatch]
-  );
-
-  React.useEffect(() => {
-    loadCollaborators(tag.id);
-  }, [tag.id, loadCollaborators]);
-
-  /**
-   * Handle remove collaborators with refresh
-   * @param tag tag
-   */
-  const handleRemoveCollaborators = (tag: GenericTag) => {
-    onRemoveCollaborators(tag, () => {
-      // Refresh collaborators after successful removal
-      loadCollaborators(tag.id);
-    });
-  };
+  const { tag, handleDelete, handleUpdate, handleRemoveCollaborators } =
+    useGuiderLabel(label);
 
   return (
     <NavigationElement
@@ -102,8 +45,8 @@ const GuiderLabel: React.FC<GuiderLabelProps> = (props: GuiderLabelProps) => {
       editableWrapperArgs={
         {
           tag,
-          onDelete,
-          onUpdate,
+          onDelete: handleDelete,
+          onUpdate: handleUpdate,
           deleteDialogTitle: t("labels.remove", {
             ns: "flags",
           }),
